make scriptEventReceiver
